# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Diary API
  description: DiaryApiIntegrationTest로 검증한 일기 API
  version: 0.0.1-SNAPSHOT
servers:
- url: /
tags: []
paths:
  /diary:
    post:
      tags:
      - Diary
      summary: 일기 생성
      description: 일기 생성
      operationId: diary-create
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/DiaryCreateRequest"
            examples:
              diary-create:
                value: "{\"title\":\"제목-{{$randomUUID}}\",\"content\":\"내용-{{$randomUUID}}\"\
                  }"
              diary-create-failed-by-blank-content:
                value: "{\"title\":\"제목-f1de13d1-1444-4aef-937f-5af6ae0579b7\",\"\
                  content\":\" \"}"
              diary-create-failed-by-long-title:
                value: "{\"title\":\"제목XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX\"\
                  ,\"content\":\"내용-f2e52cbe-3c87-4b69-a08b-8c26a9c69c1b\"}"
      responses:
        "201":
          description: "201"
          headers:
            Location:
              description: 생성된 일기 조회 경로
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diary"
              examples:
                diary-create:
                  value: "{\"diaryId\":\"01a095d9-8ac9-7c3c-a0a1-d71d20fe41b1\",\"\
                    title\":\"제목-cb4a0177-29cb-4461-a155-dc876aa4c807\",\"content\"\
                    :\"내용-6ad5365c-9076-4d50-a5dd-3541d6a2c1a4\"}"
        "400":
          description: "400"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/InvalidParameterProblem"
              examples:
                diary-create-failed-by-blank-content:
                  value: "{\"detail\":\"content: must not be blank\",\"instance\"\
                    :\"/diary\",\"status\":400,\"title\":\"Bad Request\",\"type\"\
                    :\"http://localhost:5173/problems/method-argument-not-valid\"\
                    ,\"code\":\"METHOD_ARGUMENT_NOT_VALID\"}"
                diary-create-failed-by-long-title:
                  value: "{\"detail\":\"title: size must be between 0 and 255\",\"\
                    instance\":\"/diary\",\"status\":400,\"title\":\"Bad Request\"\
                    ,\"type\":\"http://localhost:5173/problems/method-argument-not-valid\"\
                    ,\"code\":\"METHOD_ARGUMENT_NOT_VALID\"}"
  /diary/{diaryId}:
    get:
      tags:
      - Diary
      summary: 일기 조회
      description: 일기 조회
      operationId: diary-get
      parameters:
      - name: diaryId
        in: path
        description: 조회할 일기 UUID
        required: true
        schema:
          type: string
      - name: Accept-Language
        in: header
        description: "오류 설명 언어: ko는 한국어, en 및 미지원 언어는 영어"
        required: false
        schema:
          type: string
        example: ko
      responses:
        "404":
          description: "404"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              examples:
                diary-get-not-found-ko:
                  value: "{\"detail\":\"일기를 찾을 수 없습니다.\",\"instance\":\"/diary/00000000-0000-7000-8000-000000000000\"\
                    ,\"status\":404,\"title\":\"Not Found\"}"
                diary-get-not-found-en:
                  value: "{\"detail\":\"Diary not found.\",\"instance\":\"/diary/00000000-0000-7000-8000-000000000000\"\
                    ,\"status\":404,\"title\":\"Not Found\"}"
                diary-get-not-found-ja:
                  value: "{\"detail\":\"Diary not found.\",\"instance\":\"/diary/00000000-0000-7000-8000-000000000000\"\
                    ,\"status\":404,\"title\":\"Not Found\"}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diary"
              examples:
                diary-get:
                  value: "{\"diaryId\":\"01a095d9-8a5f-7991-b72f-7d8785835657\",\"\
                    title\":\"조회할 제목\",\"content\":\"조회할 내용\"}"
        "400":
          description: "400"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/InvalidParameterProblem"
              examples:
                diary-get-invalid-id:
                  value: "{\"detail\":\"Parameter ‘diaryId’ must be of type UUID.\"\
                    ,\"instance\":\"/diary/not-a-uuid\",\"status\":400,\"title\":\"\
                    Bad Request\",\"type\":\"http://localhost:5173/problems/method-argument-type-mismatch\"\
                    ,\"code\":\"METHOD_ARGUMENT_TYPE_MISMATCH\"}"
components:
  schemas:
    Diary:
      title: Diary
      required:
      - content
      - diaryId
      - title
      type: object
      properties:
        diaryId:
          type: string
          description: 서버에서 생성한 UUID v7 일기 ID
        title:
          type: string
          description: 일기 제목
        content:
          type: string
          description: 일기 내용
    InvalidParameterProblem:
      title: InvalidParameterProblem
      required:
      - code
      - detail
      - instance
      - status
      - title
      type: object
      properties:
        code:
          type: string
          description: 상세 오류 코드
        instance:
          type: string
          description: 오류가 발생한 요청 경로
        detail:
          type: string
          description: 요청 언어에 맞춘 오류 설명
        title:
          type: string
          description: HTTP 오류 제목
        type:
          type: string
          description: 문제 유형 URI (생략 시 about:blank)
          nullable: true
        status:
          type: number
          description: HTTP 상태 코드
    ProblemDetail:
      title: ProblemDetail
      required:
      - detail
      - instance
      - status
      - title
      type: object
      properties:
        instance:
          type: string
          description: 오류가 발생한 요청 경로
        detail:
          type: string
          description: 요청 언어에 맞춘 오류 설명
        title:
          type: string
          description: HTTP 오류 제목
        type:
          type: string
          description: 문제 유형 URI (생략 시 about:blank)
          nullable: true
        status:
          type: number
          description: HTTP 상태 코드
    DiaryCreateRequest:
      title: DiaryCreateRequest
      required:
      - content
      - title
      type: object
      properties:
        title:
          type: string
          description: 일기 제목
        content:
          type: string
          description: 일기 내용